requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
redis
SQLAlchemy
asyncmy
sqlmodel
python-dotenv
brapi
httpx
agno
mcp
uvloop
streamlit
streamlit-lottie
pandas
google-genai
requests